+++ main-server/src/core/identity/index.js
@@ -155,11 +155,7 @@ export async function onReceiveIdpResponseForIdentity({ nodeId, message }) {
   const responseValid = await common.getAndSaveIdpResponseValid({
     nodeId,
     requestDetail,
-    idpId: message.idp_id,
     requestDataFromMq: message,
-    responseIal: requestDetail.response_list.find(
-      (response) => response.idp_id === message.idp_id
-    ).ial,
   });
 
   let identityConsentRequestError;
@@ -196,9 +192,11 @@ export async function onReceiveIdpResponseForIdentity({ nodeId, message }) {
   }
 
   // Check if request is completed.
-  const answeredIdPCount = requestDetail.response_list.length;
+  const nonErrorAnsweredIdPCount = requestDetail.response_list.filter(
+    (response) => response.error_code == null
+  ).length;
   if (
-    answeredIdPCount === requestDetail.min_idp &&
+    nonErrorAnsweredIdPCount === requestDetail.min_idp &&
     requestStatus === 'completed'
   ) {
     let callbackFnName;